+++ template/server/actions/index.js
@@ -1,3 +1,2 @@
 module.exports.webhookHandler = () => {};
 module.exports.statusCheck = () => {};
-module.exports.updateUser = () => {};