+++ template/server/lib/sync-agent.js
@@ -1,3 +1,16 @@
-class SyncAgent {}
+// @flow
+import type { THullReqContext, THullUserUpdateMessage } from "hull";
+
+class SyncAgent {
+  init: boolean;
+
+  constructor(ctx: THullReqContext) {
+    this.init = ctx !== undefined;
+  }
+
+  sendUserUpdateMessages(messages: Array<THullUserUpdateMessage>): Promise<*> {
+    return Promise.resolve(messages);
+  }
+}
 
 module.exports = SyncAgent;