+++ template/server/server.js
@@ -1,14 +1,12 @@
 /* @flow */
 import type { $Application, $Response } from "express";
-import type { TRequest, THullReqContext } from "hull";
+import type { TRequest } from "hull";
 
-const {
-  smartNotifierHandler,
-  THullUserUpdateMessage
-} = require("hull/lib/utils");
+const { smartNotifierHandler } = require("hull/lib/utils");
 const bodyParser = require("body-parser");
 
-const { webhookHandler, statusCheck, updateUser } = require("./actions");
+const notificationsConfiguration = require("./notifications-configuration");
+const { webhookHandler, statusCheck } = require("./actions");
 
 function server(app: $Application, { token }: Object): $Application {
   app.get("/admin.html", (req: TRequest, res: $Response) => {
@@ -25,30 +23,14 @@ function server(app: $Application, { token }: Object): $Application {
       userHandlerOptions: {
         groupTraits: false
       },
-      handlers: {
-        "user:update": updateUser
-      }
+      handlers: notificationsConfiguration
     })
   );
 
   app.use(
     "/smart-notifier",
     smartNotifierHandler({
-      handlers: {
-        "user:update": (
-          ctx: THullReqContext,
-          messages: Array<THullUserUpdateMessage>
-        ) => {
-          if (ctx.smartNotifierResponse) {
-            ctx.smartNotifierResponse.setFlowControl({
-              type: "next",
-              in: parseInt(process.env.FLOW_CONTROL_IN, 10) || 1000,
-              size: parseInt(process.env.FLOW_CONTROL_SIZE, 10) || 100
-            });
-          }
-          return updateUser(ctx, messages);
-        }
-      }
+      handlers: notificationsConfiguration
     })
   );