+++ test/integration.js
@@ -10,13 +10,13 @@ const { execSync } = require("child_process");
 /**
  * Make sure we don't have anything left from previous test
  */
-console.log("cleaning up...");
+console.log(">>> cleaning up...");
 execSync("rm -rf test-connector");
 
 /**
  * Run `hull-connector-template` to generate `test-connector`
  */
-console.log("generating test-connector...");
+console.log(">>> generating test-connector...");
 execSync("node ./bin/hull-connector-template test-connector --non-interactive --name='Test Connector' --official");
 process.chdir("./test-connector");
 
@@ -25,13 +25,13 @@ process.chdir("./test-connector");
  * this step runs also build the front-end and back-end application.
  * This is what happens on Heroku deployment
  */
-console.log("installing production dependencies...");
+console.log(">>> installing production dependencies...");
 execSync("yarn --production");
 
 /**
  * Make sure the template was successfully generated
  */
-console.log("asserting generated code...");
+console.log(">>> asserting generated code...");
 const generatedPackageJson = require(path.join(process.cwd(), "package.json"));
 const generatedManifestJson = require(path.join(process.cwd(), "manifest.json"));
 
@@ -42,12 +42,12 @@ assert.equal(generatedManifestJson.name, "Test Connector");
 /**
  * Now install additional development dependencies
  */
-console.log("installing development dependencies...");
+console.log(">>> installing development dependencies...");
 execSync("yarn");
 
 
 /**
  * And run linting. `yarn test` would fail now, since we don't have full coverage in example test yet
  */
-console.log("running lint tests...");
+console.log(">>> running lint tests...");
 execSync("yarn test:lint");