+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v1.0.0
+- upgrade hull-node to v0.13.16
+- add `notifications-configuration.js` as an example of good practise
+- fix jest.config.js
+- fix base structure to encourage SyncAgent
+
 ## v0.1.2
 - upgrade `yarn` to `1.6.x`
 - apply prettier changes